requirements.txt
python-pptx>=0.6.21
streamlit>=1.42.0
